Migrate to TS and update project structure
