+++ frontend/public/index.html
@@ -109,7 +109,7 @@
     </div>
 
     <!-- Main application script -->
-    <script type="module" src="/static/js/app.js"></script>
+    <script type="module" src="../src/app.ts"></script>
 </body>
 
 </html>
