+++ frontend/src/config/configManager.ts
